Add create health-record frontend
